requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.129.2
Flask==3.1.3
fuzzywuzzy==0.18.0
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Levenshtein==0.27.3
MarkupSafe==3.0.3
openpyxl==3.1.5
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-Levenshtein==0.27.3
python-multipart==0.0.22
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1; sys_platform != "win32"
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
